% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBox-package.R
\docType{package}
\name{rotulador-package}
\alias{rotulador}
\alias{rotulador-package}
\title{rotulador: Useful Functions for Programming and Generating Documents}
\description{
Tools to help developers and producers manipulate R objects and outputs. It includes tools for displaying results and objects, and for formatting them in the correct format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TractorTom/rotulador}
  \item \url{https://tractortom.github.io/rotulador/}
  \item Report bugs at \url{https://github.com/TractorTom/rotulador/issues}
}

}
\author{
\strong{Maintainer}: Tanguy Barthelemy \email{tangbarth@hotmail.fr}

}
\keyword{internal}
