% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBox-package.R
\docType{package}
\name{TBox-package}
\alias{TBox}
\alias{TBox-package}
\title{TBox: Useful functions for programming and generating documents}
\description{
The package includes a host of tools to help developers and producers manipulate R objects and outputs. It includes tools for displaying results and objects, and for formatting them in the correct format.
}
\author{
\strong{Maintainer}: Tanguy Barthelemy \email{tanguy.barthelemy@insee.fr}

}
\keyword{internal}
