% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_code.R
\name{render_code}
\alias{render_code}
\title{Format code}
\usage{
render_code(
  output = "word",
  browser = getOption("browser"),
  eval = FALSE,
  font_size = 12
)
}
\description{
Formatter un bout de code pour le recopier dans un mail, un pdf, un document...
}
\examples{
render_code(
    output = "pdf",
    browser = "C:\\\\Program Files\\\\Adobe\\\\Adobe Illustrator CC 2015\\\\Support Files\\\\Contents\\\\Windows\\\\Illustrator.exe",
    eval = FALSE,
    font_size = 16
)


}
