% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_code.R
\name{render_code}
\alias{render_code}
\title{Generate a file with formated code}
\usage{
render_code(
  output = "word",
  browser = getOption("browser"),
  eval = FALSE,
  font_size = 12
)
}
\arguments{
\item{output}{a string. The output format}

\item{browser}{a string. The path to the browser which will open the
generated file format}

\item{eval}{a boolean specifying if the code has to be evaluated}

\item{font_size}{a numeric. The font size in pdf format.}
}
\description{
Format a piece of code to copy it into an email, a pdf, a document, etc.
}
\examples{
render_code(
    output = "pdf",
    eval = FALSE,
    font_size = 16
)

}
